pyproject.toml
[tool.poetry]
name = "mako2cli"
version = "0.1.0"
description = "Mako Template Command-Line Tool"
authors = ["fbagagli <francesco.bagagli@gmail.com>"]
license = "MIT"
readme = "README.md"
homepage = "https://github.com/fbagagli/mako2cli"
repository = "https://github.com/fbagagli/mako2cli"
keywords = ["mako", "template", "cli"]

[tool.poetry.dependencies]
python = "^3.7"
click = "^7.1.2"
pyyaml = "^5.3.1"
attrs = "^19.3.0"
mako = "^1.1.3"

[tool.poetry.dev-dependencies]
pytest = "^5.4.3"
coverage = {extras = ["toml"], version = "^5.2"}
pytest-cov = "^2.10.0"
pytest-mock = "^3.2.0"
pyfakefs = "^4.1.0"
black = "^19.10b0"
flake8 = "^3.8.3"
flake8-bandit = "^2.1.2"
flake8-black = "^0.2.1"
flake8-bugbear = "^20.1.4"
flake8-import-order = "^0.18.1"
safety = "^1.9.0"
isort = "^5.3.0"
mypy = "^0.782"
typeguard = "^2.9.1"
flake8-annotations = "^2.3.0"

[tool.poetry.scripts]
m2cli = "mako2cli.cli:main"
[tool.coverage.paths]
source = ["src", "*/site-packages"]

[tool.coverage.run]
branch = true
source = ["mako2cli"]

[tool.coverage.report]
show_missing = true
fail_under = 100

[tool.isort]
profile = "google"
src_paths = ["src", "test", "noxfile.py"]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
